+++ R/theme.R
@@ -46,11 +46,11 @@ ptclr_l <- "#e3f1ff"
 #'
 #' @param base_size Numeric text size in pts, affects all text in plot. Defaults to 11.
 #' @param gridlines Whether to display major gridlines along `"y"` (the default),
-#' `"x"`, `"both"` or draw a `"scatter"`, which has both gridlines and inverted colours.
+#' `"x"`, `"both"`, or draw a `"scatter"`, which has both gridlines and inverted colours.
 #' @param family,title_family font family to use for the (title of the) plot. Defaults to `"IBM Plex Sans"` for title and `"IBM Plex Sans Condensed"` for plot.
 #' @param margin_side,margin_bottom size of left and right / bottom margin around plot, in pts. Defaults to 6. Set to 0 to align flush with text in a Word document.
 #' @param plot.title.position where to align the title. Either "plot" (the default, difference from `theme()` default) or `"panel"`.
-#' @param axis.title same as in theme(), but with a default of `element_blank()`.
+#' @param axis_titles whether to display axis titles: `"none"` (default), `"x"`, `"y"`, or `"both"`.
 #' @param multiplot if set to TRUE, provides better styling for small multiples created using `facet_*`.
 #' @param map if set to TRUE, provides better styling for maps created using `geom_sf()`. Overrides `gridlines`.
 #' @inheritDotParams ggplot2::theme -text -plot.title -panel.grid.minor
@@ -118,26 +118,81 @@ theme_ptrr <- function(gridlines = c("y", "x", "both", "scatter", "none"),
                          margin_side = 6,
                          margin_bottom = 6,
                          plot.title.position = "plot",
-                         axis_titles = FALSE,
+                         axis_titles = c("none", "y", "x", "both"),
                          richtext = FALSE,
+                         richtext_style = NULL,
                          map = FALSE,
+                         inverse = FALSE,
                          ...) {
 
   if(richtext) {
-    if(!requireNamespace("ggtext")) stop("Package ggtext needed.")
+    if(!requireNamespace("marquee")) stop("Package marquee needed.")
   }
 
-  element_switch <- if(richtext) ggtext::element_markdown else ggplot2::element_text
+  element_switch <- if (!richtext) {
+    function(...) {
+      funArgs <-  list(...)
+      # remove style argument if not using richtext, dont pass it to element_text
+      newArgs <- funArgs[is.na(match(names(funArgs), "style"))]
+      do.call(ggplot2::element_text, newArgs)
+    }
+  }  else {
+    function(...) {
+      funArgs <-  list(...)
+      do.call(marquee::element_marquee, funArgs)
+    }
+  }
 
   grd <- match.arg(gridlines)
-  grid_col <- if(grd == "scatter" | multiplot) "white" else "grey92"
+  if(axis_titles == TRUE) axis_titles <- "both"
+  if(axis_titles == FALSE) axis_titles <- "none"
+  axttl <- match.arg(axis_titles)
+
+  grid_col <- if(grd == "scatter" | multiplot) "white" else tonecol
   bg_col <- if(grd == "scatter" | multiplot) tonecol else "white"
+
+  if(!inverse & grd != "scatter" & !multiplot) {
+    plot_bg_col <- "white"
+    bg_col <- tonecol
+    grid_col <- "white"
+  } else if (inverse & (grd == "scatter" | multiplot)) {
+    plot_bg_col <- tonecol
+    bg_col <- "white"
+    grid_col <- tonecol
+  } else if (inverse == TRUE) {
+    plot_bg_col <- tonecol
+    bg_col <- "white"
+    grid_col <- tonecol
+  } else if (!inverse == TRUE & grd == "scatter") {
+    plot_bg_col <- "white"
+    bg_col <- tonecol
+    grid_col <- "white"
+  } else {
+    plot_bg_col <- "white"
+    bg_col <- "white"
+    grid_col <- tonecol
+  }
+
+  if(is.null(richtext_style)) {
+    richtext_style <- marquee::classic_style() |>
+      marquee::modify_style(tag = "body", lineheight = 1,
+                            margin = marquee::trbl(0, 0, 0, 0))
+  }
+
+  plt_title <- if(richtext) {
+
+    marquee::element_marquee(size = base_size * 1.2, family = title_family,
+                             style = richtext_style |>
+                               marquee::modify_style(tag = "body", weight = 700))
+  } else {
+    ggplot2::element_text(face = "bold", size = base_size * 1.2,
+                          family = title_family)
+  }
+
   element_gridline <- ggplot2::element_line(colour = grid_col, size = 0.3)
   thm <- ggplot2::theme_minimal(base_size = base_size, base_family = family) +
     ggplot2::theme(plot.title.position = plot.title.position,
-                   plot.title = element_switch(face = "bold",
-                                                      size = base_size * 1.2,
-                                                      family = title_family),
+                   plot.title = plt_title,
                    plot.subtitle = element_switch(),
                    plot.caption = element_switch(colour = "grey60"),
                    panel.grid.minor = ggplot2::element_blank(),
@@ -148,17 +203,25 @@ theme_ptrr <- function(gridlines = c("y", "x", "both", "scatter", "none"),
                    # axis.line = ggplot2::element_line(),
                    panel.background = ggplot2::element_rect(fill = bg_col,
                                                             colour = NA),
-                   axis.title = if(axis_titles) element_switch() else ggplot2::element_blank(),
-                   strip.text.x = element_switch(hjust = 0),
-                   strip.text.y = element_switch(hjust = 0),
-                   axis.text.x = element_switch(),
-                   axis.text.y = element_switch(),
-                   plot.margin = ggplot2::unit(c(10, margin_side,
+                   axis.title = if(axttl == "both" | axttl == TRUE) element_switch() else ggplot2::element_blank(),
+                   axis.title.x = if(axttl == "both" | axttl == "x") element_switch() else ggplot2::element_blank(),
+                   axis.title.y = if(axttl == "both" | axttl == "y") element_switch() else ggplot2::element_blank(),
+                   strip.text.x = element_switch(hjust = 0, style = richtext_style),
+                   strip.text.y = element_switch(hjust = 0, style = richtext_style),
+                   plot.background = element_rect(fill = plot_bg_col, colour = NA),
+                   axis.text.x = element_switch(style = richtext_style),
+                   axis.text.y = element_switch(style = richtext_style),
+                   plot.margin = ggplot2::unit(c(10, margin_side + 5,
                                                  margin_bottom, margin_side),
                                                units = "pt"))
-  if(multiplot) thm <- thm +
+  if(multiplot & !inverse) {thm <- thm +
     ggplot2::theme(strip.background = ggplot2::element_rect(fill = tonecol,
                                                             colour = NA))
+  } else if(multiplot & inverse) {
+    thm <- thm +
+      ggplot2::theme(strip.background = ggplot2::element_rect(fill = "white",
+                                                              colour = NA))
+  }
 
   if(map) thm <- thm +
     ggplot2::theme(panel.grid.major = ggplot2::element_blank(),